api/requirements.txt
tensorflow
keras
Pillow
scikit-image
aiofiles
python-multipart
